Pull dependencies via web, create asset manifest
* Removes git submodule for mozart dependencies
* Adds new :clean rake task to remove old dependencies
* Adds new :download rake task to pull JavaScript files from web
* Removes :copy_source_files rake task
* Adds new class file for assets related constants (not ideal)
* Adds new mozart-rails asset manifest for cleaner and tidier usage
* Adds GEM_ROOT constant
* Removes JavaScript file dependencies and replaces them with gem versions
* Updates README
* Bumps gem version to 1.0.1

+++ Rakefile
@@ -1,10 +1,12 @@
 require "bundler/gem_tasks"
+require "mozart/rails/assets"
 
-task :release => [:clean, :copy_source_files, :guard_version]
+task :build => [:clean, :download]
+task :release => [:clean, :download, :guard_version]
 
 task :guard_version do
   def check_version(file, pattern, constant)
-    body = File.read("vendor/assets/javascripts/#{file}")
+    body = File.read(Mozart::Rails::VENDOR_PATH + '/' + file)
     match = body.match(pattern) or abort "Version check failed: no pattern matched in #{file}"
     file_version = match[1]
     constant_version = Mozart::Rails.const_get(constant)
@@ -21,29 +23,40 @@ desc "Remove JavaScript asset files"
 task :clean do
   puts "Removing JavaScript assets"
 
-  Dir.glob("vendor/assets/javascripts/*.js").each do |file|
-    puts "Removing #{file}"
-    File.unlink file
+  Dir.glob(Mozart::Rails::VENDOR_PATH + "/*.js").each do |file|
+    unless file.match(/mozart-rails.js$/)
+      puts "Removing #{file}"
+      File.unlink file
+    end
   end
 end
 
-desc "Copy source files into the assets directory"
-task :copy_source_files do
-  require 'fileutils'
+desc "Download source files into the assets directory"
+task :download do
+  require 'net/https'
 
-  VENDOR_PATH = "vendor/mozart-empty/app/assets/js"
-  ASSETS_PATH = "vendor/assets/javascripts"
+  Mozart::Rails::JAVASCRIPT_URLS.each do |key, urls|
+    urls.each do |url|
+      uri = URI.parse(url)
 
-  [
-    "handlebars-#{Mozart::Rails::HANDLEBARS_VERSION}",
-    "jquery-#{Mozart::Rails::JQUERY_VERSION}",
-    "mozart.min",
-    "underscore-#{Mozart::Rails::UNDERSCORE_VERSION}",
-  ].each do |file|
-    if File.file? File.join(ASSETS_PATH, "#{file}.js")
-      puts "File exists: #{File.join(ASSETS_PATH, "#{file.gsub(/-.+/, '')}.js")}, skipping"
-    else
-      FileUtils.cp File.join(VENDOR_PATH, "#{file}.js"), File.join(ASSETS_PATH, "#{file.gsub(/-.+/, '')}.js")
+      Net::HTTP.start(uri.host, uri.port, :use_ssl => uri.scheme == 'https') do |http|
+        request = Net::HTTP::Get.new uri.path
+        response = http.request request
+
+        case response
+        when Net::HTTPSuccess then
+          ext = url.match(/(min).js$/) ? '.min.js' : '.js'
+          dest = File.join(Mozart::Rails::VENDOR_PATH, key.to_s + ext)
+
+          puts "Downloading #{url} to #{dest}"
+
+          File.open(dest, 'w+') do |f|
+            f.puts response.body
+          end
+        else
+          raise "Unexpected response: #{response.inspect}"
+        end
+      end
     end
   end
 end
